Core and GUI implementation

+++ capy/db/migrate/20160813132504_create_pages.rb
@@ -1,8 +1,8 @@
 class CreatePages < ActiveRecord::Migration[5.0]
   def change
     create_table :pages do |t|
-      t.integer :last_date
-      t.integer :last_duration
+      t.bigint :last_date
+      t.bigint :last_duration
       t.string :browser_name
       t.string :browser_version
       t.string :url
